Move _wait_until_in_progress to buildercore.cloudformation
